requirements.txt
click
PyYAML
dacite